Added common code to one place and fixed other review comments
