+++ examples/spark/src/main/scala/org/apache/carbondata/examples/util/ExampleUtils.scala
@@ -85,6 +85,7 @@ object ExampleUtils {
     } else {
       "local[" + workThreadNum.toString() + "]"
     }
+    // TODO: Analyse the legacy configs and add test cases for non legacy ones
     val spark = SparkSession
       .builder()
       .master(masterUrl)
