+++ integration/spark/src/main/common2.3and2.4/org/apache/spark/sql/hive/CarbonAnalyzer.scala
