+++ integration/spark/src/main/scala/org/apache/spark/sql/execution/strategy/CarbonDataSourceScan.scala
@@ -20,27 +20,19 @@ import scala.collection.JavaConverters._
 
 import org.apache.spark.CarbonInputMetrics
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{CarbonDatasourceHadoopRelation, SparkVersionAdapter}
+import org.apache.spark.sql.{CarbonDatasourceHadoopRelation, CarbonDataSourceScanHelper}
 import org.apache.spark.sql.carbondata.execution.datasources.CarbonSparkDataSourceUtil
 import org.apache.spark.sql.catalyst.{InternalRow, TableIdentifier}
 import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeReference, SortOrder, UnsafeProjection}
 import org.apache.spark.sql.catalyst.expressions.{Expression => SparkExpression}
-import org.apache.spark.sql.catalyst.expressions.codegen.CodegenContext
 import org.apache.spark.sql.catalyst.plans.QueryPlan
 import org.apache.spark.sql.catalyst.plans.physical.{HashPartitioning, Partitioning, UnknownPartitioning}
-import org.apache.spark.sql.execution.{DataSourceScanExec, WholeStageCodegenExec}
-import org.apache.spark.sql.execution.metric.SQLMetrics
-import org.apache.spark.sql.optimizer.CarbonFilters
-import org.apache.spark.sql.vectorized.ColumnarBatch
+import org.apache.spark.sql.execution.WholeStageCodegenExec
 
-import org.apache.carbondata.core.index.IndexFilter
-import org.apache.carbondata.core.indexstore.PartitionSpec
 import org.apache.carbondata.core.metadata.schema.BucketingInfo
 import org.apache.carbondata.core.readcommitter.ReadCommittedScope
 import org.apache.carbondata.core.scan.expression.Expression
-import org.apache.carbondata.core.scan.expression.logical.AndExpression
 import org.apache.carbondata.hadoop.CarbonProjection
-import org.apache.carbondata.spark.rdd.CarbonScanRDD
 
 /**
  *  Physical plan node for scanning data. It is applied for both tables
@@ -51,41 +43,15 @@ case class CarbonDataSourceScan(
     output: Seq[Attribute],
     partitionFilters: Seq[SparkExpression],
     dataFilters: Seq[SparkExpression],
-    @transient readComittedScope: ReadCommittedScope,
+    @transient readCommittedScope: ReadCommittedScope,
     @transient pushedDownProjection: CarbonProjection,
     @transient pushedDownFilters: Seq[Expression],
     directScanSupport: Boolean,
     @transient extraRDD: Option[(RDD[InternalRow], Boolean)] = None,
     tableIdentifier: Option[TableIdentifier] = None,
     segmentIds: Option[String] = None)
-  extends DataSourceScanExec {
-
-  override lazy val supportsColumnar: Boolean = CarbonPlanHelper
-    .supportBatchedDataSource(sqlContext, output, extraRDD)
-
-  override protected def doExecuteColumnar(): RDD[ColumnarBatch] = {
-    val numOutputRows = longMetric("numOutputRows")
-    inputRDD.asInstanceOf[RDD[ColumnarBatch]].mapPartitionsInternal { batches =>
-      new Iterator[ColumnarBatch] {
-
-        override def hasNext: Boolean = {
-          val res = batches.hasNext
-          res
-        }
-
-        override def next(): ColumnarBatch = {
-          val batch = batches.next()
-          numOutputRows += batch.numRows()
-          batch
-        }
-      }
-    }
-  }
-
-  override lazy val metrics = Map(
-    "numOutputRows" -> SQLMetrics.createMetric(sparkContext, "number of output rows"))
-
-  lazy val needsUnsafeRowConversion: Boolean = { true }
+  extends CarbonDataSourceScanHelper(relation, output, partitionFilters, pushedDownFilters,
+    pushedDownProjection, directScanSupport, extraRDD, segmentIds) {
 
   override lazy val (outputPartitioning, outputOrdering): (Partitioning, Seq[SortOrder]) = {
     val info: BucketingInfo = relation.carbonTable.getBucketingInfo
@@ -119,8 +85,8 @@ case class CarbonDataSourceScan(
     val metadata =
       Map(
         "ReadSchema" -> seqToString(pushedDownProjection.getAllColumns),
-        "Batched" -> supportsColumnar.toString,
-        "DirectScan" -> (supportsColumnar && directScanSupport).toString,
+        "Batched" -> supportsBatchOrColumnar.toString,
+        "DirectScan" -> (supportsBatchOrColumnar && directScanSupport).toString,
         "PushedFilters" -> seqToString(pushedDownFilters.map(_.getStatement)))
     if (relation.carbonTable.isHivePartitionTable) {
       metadata + ("PartitionFilters" -> seqToString(partitionFilters)) +
@@ -130,42 +96,10 @@ case class CarbonDataSourceScan(
     }
   }
 
-  @transient private lazy val indexFilter: IndexFilter = {
-    val filter = pushedDownFilters.reduceOption(new AndExpression(_, _))
-      .map(new IndexFilter(relation.carbonTable, _, true)).orNull
-    if (filter != null && pushedDownFilters.length == 1) {
-      // push down the limit if only one filter
-      filter.setLimit(relation.limit)
-    }
-    filter
-  }
-
-  @transient private lazy val selectedPartitions: Seq[PartitionSpec] = {
-    CarbonFilters
-      .getPartitions(partitionFilters, relation.sparkSession, relation.carbonTable)
-      .orNull
-  }
-
-  private lazy val inputRDD: RDD[InternalRow] = {
-    val carbonRdd = new CarbonScanRDD[InternalRow](
-      relation.sparkSession,
-      pushedDownProjection,
-      indexFilter,
-      relation.identifier,
-      relation.carbonTable.getTableInfo.serialize(),
-      relation.carbonTable.getTableInfo,
-      new CarbonInputMetrics,
-      selectedPartitions,
-      segmentIds = segmentIds)
-    carbonRdd.setVectorReaderSupport(supportsColumnar)
-    carbonRdd.setDirectScanSupport(supportsColumnar && directScanSupport)
-    extraRDD.map(_._1.union(carbonRdd)).getOrElse(carbonRdd)
-  }
-
   override def inputRDDs(): Seq[RDD[InternalRow]] = inputRDD :: Nil
 
   override protected def doExecute(): RDD[InternalRow] = {
-    if (supportsColumnar) {
+    if (supportsBatchOrColumnar) {
       // in the case of fallback, this batched scan should never fail because of:
       // 1) only primitive types are supported
       // 2) the number of columns should be smaller than spark.sql.codegen.maxFields
@@ -194,7 +128,7 @@ case class CarbonDataSourceScan(
   override protected def doCanonicalize(): CarbonDataSourceScan = {
     CarbonDataSourceScan(
       relation,
-      output.map(SparkVersionAdapter.normalizeExpressions(_, output)),
+      a,
       QueryPlan.normalizePredicates(partitionFilters, output),
       QueryPlan.normalizePredicates(dataFilters, output),
       null,
@@ -203,9 +137,5 @@ case class CarbonDataSourceScan(
       directScanSupport,
       extraRDD,
       tableIdentifier)
-  }
-
-  protected def doProduce(ctx: CodegenContext): String = {
-    WholeStageCodegenExec(this)(1).doProduce(ctx)
   }
 }
