+++ integration/spark/src/main/spark2.4/org/apache/spark/sql/CarbonToSparkAdapter.scala
@@ -155,8 +155,7 @@ object CarbonToSparkAdapter {
       name: String,
       exprId: ExprId = NamedExpression.newExprId,
       qualifier: Seq[String] = Seq.empty,
-      explicitMetadata: Option[Metadata] = None,
-      namedExpr: Option[NamedExpression] = None) : Alias = {
+      explicitMetadata: Option[Metadata] = None) : Alias = {
     Alias(child, name)(exprId, qualifier, explicitMetadata)
   }
