--- integration/spark/src/main/spark3.1/org/apache/spark/sql/execution/CreateDataSourceTableCommand.scala
@@ -1,59 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.spark.sql.execution
-
-import org.apache.log4j.Logger
-import org.apache.spark.sql.{AnalysisException, Row, SparkSession}
-import org.apache.spark.sql.catalyst.catalog.{CatalogTable, CatalogTableType, CatalogUtils}
-import org.apache.spark.sql.execution.command.RunnableCommand
-import org.apache.spark.sql.util.CreateTableCommonUtil.getNewTable
-
-import org.apache.carbondata.common.logging.LogServiceFactory
-
-case class CreateDataSourceTableCommand(table: CatalogTable, ignoreIfExists: Boolean)
-  extends RunnableCommand {
-
-  val LOGGER: Logger = LogServiceFactory.getLogService(this.getClass.getName)
-
-  override def run(sparkSession: SparkSession): Seq[Row] = {
-    assert(table.tableType != CatalogTableType.VIEW)
-    assert(table.provider.isDefined)
-
-    val sessionState = sparkSession.sessionState
-    if (sessionState.catalog.tableExists(table.identifier)) {
-      if (ignoreIfExists) {
-        return Seq.empty[Row]
-      } else {
-        throw new AnalysisException(s"Table ${table.identifier.unquotedString} already exists.")
-      }
-    }
-    val newTable: CatalogTable = getNewTable(sparkSession, sessionState, table, LOGGER)
-
-    // We will return Nil or throw exception at the beginning if the table already exists, so when
-    // we reach here, the table should not exist and we should set `ignoreIfExists` to false.
-    sessionState.catalog.createTable(newTable, ignoreIfExists = false, false)
-    Seq.empty[Row]
-  }
-}
-
-object CreateDataSourceTableCommand {
-  def createDataSource(catalogTable: CatalogTable,
-      ignoreIfExists: Boolean,
-      sparkSession: SparkSession): Seq[Row] = {
-    CreateDataSourceTableCommand(catalogTable, ignoreIfExists).run(sparkSession)
-  }
-}
