+++ integration/spark/src/main/spark3.1/org/apache/spark/sql/hive/execution/command/CarbonResetCommand.scala
@@ -17,7 +17,7 @@
 
 package org.apache.spark.sql.hive.execution.command
 
-import org.apache.spark.sql.{CarbonEnv, CarbonToSparkAdapter, Row, SparkSession}
+import org.apache.spark.sql.{CarbonEnv, Row, SparkSession}
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.execution.command.{ResetCommand, RunnableCommand}
