+++ mv/plan/src/main/scala/org/apache/carbondata/mv/plans/modular/Harmonizer.scala
@@ -86,7 +86,7 @@ object HarmonizeDimensionTable extends Rule[ModularPlan] with PredicateHelper {
                 case a: Alias  if a.child.isInstanceOf[Attribute] =>
                 (a.child.asInstanceOf[Attribute], a.toAttribute) })
             val pullUpPredicates = s1.predicateList
-              .map(replaceAlias(_, ExpressionHelper.attributeMap(rAliasMap)))
+              .map(replaceAlias(_, SparkVersionHelper.attributeMap(rAliasMap)))
             if (pullUpPredicates.forall(cond => canEvaluate(cond, h))) {
               tPullUpPredicates = tPullUpPredicates ++ pullUpPredicates
               plans.modular.HarmonizedRelation(h.copy(child = s1.copy(predicateList = Nil)))
